+++ binfmt/elf.c
@@ -69,11 +69,6 @@ static int elf_loadbinary(FAR struct binary_s *binp,
                           FAR const char *filename,
                           FAR const struct symtab_s *exports,
                           int nexports);
-#ifdef CONFIG_ELF_COREDUMP
-static int elf_dumpbinary(FAR const struct memory_region_s *regions,
-                          FAR struct lib_outstream_s *stream,
-                          pid_t pid);
-#endif
 #if defined(CONFIG_DEBUG_FEATURES) && defined(CONFIG_DEBUG_BINFMT)
 static void elf_dumploadinfo(FAR struct elf_loadinfo_s *loadinfo);
 #endif
@@ -87,9 +82,6 @@ static struct binfmt_s g_elfbinfmt =
   NULL,             /* next */
   elf_loadbinary,   /* load */
   NULL,             /* unload */
-#ifdef CONFIG_ELF_COREDUMP
-  elf_dumpbinary,   /* coredump */
-#endif
 };
 
 /****************************************************************************
@@ -369,32 +361,6 @@ errout_with_init:
   elf_uninit(&loadinfo);
   return ret;
 }
-
-/****************************************************************************
- * Name: elf_dumpbinary
- *
- * Description:
- *   Generat the core dump stream as ELF structure.
- *
- * Returned Value:
- *   Zero (OK) on success; a negated errno value on failure.
- *
- ****************************************************************************/
-
-#ifdef CONFIG_ELF_COREDUMP
-static int elf_dumpbinary(FAR const struct memory_region_s *regions,
-                          FAR struct lib_outstream_s *stream,
-                          pid_t pid)
-{
-  struct elf_dumpinfo_s dumpinfo;
-
-  dumpinfo.regions = regions;
-  dumpinfo.stream  = stream;
-  dumpinfo.pid     = pid;
-
-  return elf_coredump(&dumpinfo);
-}
-#endif
 
 /****************************************************************************
  * Public Functions
