+++ binfmt/libelf/Make.defs
@@ -28,12 +28,6 @@ CSRCS += libelf_bind.c libelf_init.c libelf_addrenv.c libelf_iobuffer.c
 CSRCS += libelf_load.c libelf_read.c libelf_sections.c libelf_symbols.c
 CSRCS += libelf_uninit.c libelf_unload.c libelf_verify.c
 
-ifeq ($(CONFIG_ELF_COREDUMP),y)
-CSRCS += libelf_coredump.c
-
-CFLAGS += ${INCDIR_PREFIX}$(TOPDIR)$(DELIM)sched
-endif
-
 ifeq ($(CONFIG_BINFMT_CONSTRUCTORS),y)
 CSRCS += libelf_ctors.c libelf_dtors.c
 endif
