+++ boards/arm/imx6/sabre-6quad/configs/coredump/defconfig
@@ -21,13 +21,12 @@ CONFIG_BOARD_COREDUMP_SYSLOG=y
 CONFIG_BOARD_LOOPSPERMSEC=99369
 CONFIG_BOOT_RUNFROMSDRAM=y
 CONFIG_BUILTIN=y
+CONFIG_COREDUMP=y
 CONFIG_DEBUG_ASSERTIONS=y
 CONFIG_DEBUG_FEATURES=y
 CONFIG_DEBUG_FULLOPT=y
 CONFIG_DEBUG_SYMBOLS=y
 CONFIG_DEV_ZERO=y
-CONFIG_ELF=y
-CONFIG_ELF_COREDUMP=y
 CONFIG_EXAMPLES_HELLO=y
 CONFIG_EXPERIMENTAL=y
 CONFIG_FS_PROCFS=y
