+++ include/nuttx/binfmt/binfmt.h
@@ -31,18 +31,14 @@
 
 #include <spawn.h>
 #include <sys/types.h>
-#include <sys/utsname.h>
 
 #include <nuttx/sched.h>
-#include <nuttx/streams.h>
-#include <nuttx/memoryregion.h>
 
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
 
 #define BINFMT_NALLOC     4
-#define COREDUMP_MAGIC    0x434f5245
 
 /****************************************************************************
  * Public Types
@@ -138,22 +134,6 @@ struct binfmt_s
   /* Unload module callback */
 
   CODE int (*unload)(FAR struct binary_s *bin);
-
-  /* Coredump callback */
-
-  CODE int (*coredump)(FAR const struct memory_region_s *regions,
-                       FAR struct lib_outstream_s *stream,
-                       pid_t pid);
-};
-
-/* Coredump information for block header */
-
-struct coredump_info_s
-{
-  uint32_t       magic;
-  struct utsname name;
-  time_t         time;
-  size_t         size;
 };
 
 /****************************************************************************
@@ -208,23 +188,6 @@ int register_binfmt(FAR struct binfmt_s *binfmt);
  ****************************************************************************/
 
 int unregister_binfmt(FAR struct binfmt_s *binfmt);
-
-/****************************************************************************
- * Name: core_dump
- *
- * Description:
- *   This function for generating core dump stream.
- *
- * Returned Value:
- *   This is a NuttX internal function so it follows the convention that
- *   0 (OK) is returned on success and a negated errno is returned on
- *   failure.
- *
- ****************************************************************************/
-
-int core_dump(FAR const struct memory_region_s *regions,
-              FAR struct lib_outstream_s *stream,
-              pid_t pid);
 
 /****************************************************************************
  * Name: load_module
