+++ include/nuttx/elf.h
@@ -40,7 +40,7 @@
  * Public Types
  ****************************************************************************/
 
-#ifdef CONFIG_ELF_COREDUMP
+#ifdef CONFIG_COREDUMP
 typedef struct elf_prpsinfo_s
 {
   char           pr_state;    /* Numeric process state */
