+++ sched/misc/CMakeLists.txt
@@ -26,7 +26,7 @@ if(CONFIG_ARCH_DEADLOCKDUMP)
   list(APPEND SRCS deadlock.c)
 endif()
 
-if(CONFIG_BOARD_COREDUMP_SYSLOG OR CONFIG_BOARD_COREDUMP_BLKDEV)
+if(CONFIG_COREDUMP)
   list(APPEND SRCS coredump.c)
 endif()
